blake2: remove SIMD support
