+++ blake2/Cargo.toml
@@ -24,9 +24,6 @@ default = ["alloc"]
 alloc = ["digest/alloc"]
 zeroize = ["digest/zeroize"]
 reset = [] # Enable reset functionality
-#simd = []
-#simd_opt = ["simd"]
-#simd_asm = ["simd_opt"]
 size_opt = [] # Optimize for code size. Removes some `inline(always)`
 
 [package.metadata.docs.rs]
