+++ blake2/src/consts.rs
@@ -1,5 +1,3 @@
-#![allow(clippy::unreadable_literal)]
-
 pub(crate) static SIGMA: [[usize; 16]; 12] = [
     [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],
     [14, 10, 4, 8, 9, 15, 13, 6, 1, 12, 0, 2, 11, 7, 5, 3],
@@ -26,22 +24,6 @@ pub(crate) static BLAKE2B_IV: [u64; 8] = [
     0x5be0cd19137e2179,
 ];
 
-/*
-pub const BLAKE2B_BLOCKBYTES : usize = 128;
-pub const BLAKE2B_OUTBYTES : usize = 64;
-pub const BLAKE2B_KEYBYTES : usize = 64;
-pub const BLAKE2B_SALTBYTES : usize = 16;
-pub const BLAKE2B_PERSONALBYTES : usize = 16;
-*/
-
 pub(crate) static BLAKE2S_IV: [u32; 8] = [
     0x6A09E667, 0xBB67AE85, 0x3C6EF372, 0xA54FF53A, 0x510E527F, 0x9B05688C, 0x1F83D9AB, 0x5BE0CD19,
 ];
-
-/*
-pub const BLAKE2S_BLOCKBYTES : usize = 64;
-pub const BLAKE2S_OUTBYTES : usize = 32;
-pub const BLAKE2S_KEYBYTES : usize = 32;
-pub const BLAKE2S_SALTBYTES : usize = 8;
-pub const BLAKE2S_PERSONALBYTES : usize = 8;
-*/
